Added comments and X/O placement
Added a bunch of comments to make it easier for you to figure out what each thing does. Also made it to where the X and O are placed when the user clicks. There's still a lot of logic left (you can override squares, and we never check for a winner).