Fixed a probable typo that set the squares filled by O to -1.
Added a host of helper functions to help with indexing between 1d and 2d arrays as well as do some vectorized math (multiplication and addition)

Added win conditions for row and columns only. Still need to implement diagonal checks. Winner is printed to console for now, will probably want a pop-up window.
Winner is determined by the sum of the most recently placed square's row and column data. Since X is 0 and O is 1, any rows that sum to 0 are filled with X, and any that sum to 3 are filled with O. NaN as the default value in each square keeps the game from ending prior to filling the entire row/column with something (since NaN plus anything is NaN)